% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca-seurat.R
\name{pca_seurat}
\alias{pca_seurat}
\title{PCA on Seurat Data}
\usage{
pca_seurat(input)
}
\arguments{
\item{input}{is the file after scaling}
}
\description{
This package performs linear dimensional reduction on the seurat object, generating png files
}
\examples{
pca_seurat("after_scaling.rds")
}
